requirements.txt
python-dotenv==1.0.1
requests==2.28.2
tweepy==4.14.0
anthropic==0.49.0